Pacman

+++ inc/IDisplayModule.hpp
@@ -29,8 +29,8 @@ namespace arc
             virtual bool isOk() = 0;
             virtual bool getKeyDown(arc::Keyboard key) = 0;
 
-            virtual void drawText(const std::string& text, int characterSize, Color color, std::pair<int, int> position) = 0;
-            virtual void drawSquare(int size, Color color, std::pair<int, int> position) = 0;
+            virtual void drawText(const std::string& text, int characterSize, Color color, std::pair<float, float> position) = 0;
+            virtual void drawSquare(int size, Color color, std::pair<float, float> position) = 0;
 
     }; // class IDisplayModule
